Freeze git URLs properly: save the revision ID rather than pkg version
The default behavior was to save a requirement referenced through a git
URL as a standard "pkg_name==version" dependency. This would cause pip
to try to install the dependency from pypi when using the requirements
file, which is usually not the same version.

Instead, get the revision ID of the source currently checked out by pip
and save that in the requirements file. This ensures that future runs of
pip will always get the exact same code.